+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.repository;
 
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
