+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.repository;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
