+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.repository;
 
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
