+++ src/main/java/kitchenpos/table/repository/OrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.table.domain;
+package kitchenpos.table.repository;
 
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
