+++ src/main/java/kitchenpos/table/repository/TableGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.table.domain;
+package kitchenpos.table.repository;
 
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
